checkPointProtecaoBanco/requirements .txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu

-f https://dl.fbaipublicfiles.com/detectron2/wheels/cu101/torch1.7/index.html
detectron2==0.5+cu101

captcha
pyyaml==5.1
django_recaptcha==3.0.0
numpy==1.21.6
opencv_python==4.6.0.66
pandas==1.3.5
Pillow==9.2.0
protobuf==4.21.5
streamlit==1.12.2
